(feat) O3-4023 Ward App - force patient to be admitted before being able to transfer
