+++ packages/esm-ward-app/src/ward-view-header/admission-requests.scss
@@ -36,8 +36,8 @@
 }
 
 .lightBlueBackground {
-  background-color: white;
-  border-left:4px solid $color-blue-60-2;
+  background-color: $color-blue-10;
+  border-left: 4px solid $color-blue-60-2;
   color: #393939;
 }
