+++ packages/esm-ward-app/src/ward-workspace/patient-transfer-bed-swap/patient-transfer-swap.scss
@@ -9,6 +9,10 @@
   color: #393939;
 }
 
+.formError {
+  margin-bottom: layout.$spacing-05;
+}
+
 .workspaceContent {
   padding: layout.$spacing-05;
   height: 100%;
